refactor(storage): align storage/rdb package layout with model layer
- Move storage/rdb/cluster_conf/api_key.go -> storage/rdb/api_key/api_key.go.
- Move storage/rdb/quota/entity*.go -> storage/rdb/entity/.
- Move storage/rdb/quota/rate_limit_policy.go -> storage/rdb/rate_limit_policy/rate_limit_policy.go.
- Move storage/rdb/quota/route_rules.go -> storage/rdb/route_rules/route_rules.go.
- Update stateful/container/rdb/components.go imports and constructor calls with package aliases.
- Fix innerapi_v1 test stubs using wrong model packages (entity vs rate_limit_policy / quota).
- Update sys-design docs to reflect new storage layout.

All unit tests and model_price integration tests pass.
