+++ design-docs/sys-design/details/限流策略与导出.md
@@ -287,7 +287,7 @@ BFE 收到配置后：
 |------|------|
 | `model/rate_limit_policy/rate_limit_policy.go` | 限流策略业务模型、JSON 序列化、BFE 导出结构 |
 | `model/rate_limit_policy/rate_limit_policy_manager.go` | CRUD、校验、导出生成器 |
-| `storage/rdb/quota/rate_limit_policy.go` | 数据库实体与查询 |
+| `storage/rdb/rate_limit_policy/rate_limit_policy.go` | 数据库实体与查询 |
 | `model/imods/exporter.go` | 统一导出调度 |
 | `model/api_key/api_key.go` | API-Key 与限流策略关联字段 |
 | `model/entity/entity.go` / `entity_manager.go` | Entity 与限流策略关联字段 |
