+++ design-docs/sys-design/总体设计文档.md
@@ -259,10 +259,14 @@ ai-gateway-api/
 |--------|------------|--------|
 | `storage/rdb/auth` | `model/iauth` | `users`、`user_products` |
 | `storage/rdb/basic` | `model/ibasic` | `bfe_clusters`、`products`、`extra_files` |
-| `storage/rdb/cluster_conf` | `model/icluster_conf`、`model/api_key` | `clusters`、`sub_clusters`、`pools`、`api_keys`、`lb_matrices` |
+| `storage/rdb/api_key` | `model/api_key` | `api_keys`、`api_key_tokens` |
+| `storage/rdb/cluster_conf` | `model/icluster_conf` | `clusters`、`sub_clusters`、`pools`、`lb_matrices` |
+| `storage/rdb/entity` | `model/entity` | `entity_types`、`entities` |
 | `storage/rdb/protocol` | `model/iprotocol` | `certificates` |
-| `storage/rdb/quota` | `model/entity`、`model/rate_limit_policy`、`model/quota`、`model/shared` | `entity_types`、`entities`、`quota_plans`、`quota_balances`、`rate_limit_policies`、`route_rules` |
+| `storage/rdb/quota` | `model/quota` | `quota_plans`、`quota_balances` |
+| `storage/rdb/rate_limit_policy` | `model/rate_limit_policy` | `rate_limit_policies` |
 | `storage/rdb/route_conf` | `model/iroute_conf` | `domains`、`route_basic_rules`、`route_advance_rules`、`route_default_rules`、`route_cases` |
+| `storage/rdb/route_rules` | `model/shared`、`model/route_rules` | `route_rules` |
 | `storage/rdb/ai_route` | `model/iai_route` | `ai_route_rules` |
 | `storage/rdb/version_control` | `model/iversion_control` | `config_versions` |
