+++ design-docs/sys-design/模型层设计文档.md
@@ -2079,11 +2079,15 @@ type TxnStorager interface {
 |------|------|------|
 | `basic` | `product.go`、`bfe_cluster.go`、`extra_file.go` | 基础资源存储 |
 | `auth` | `authentication.go`、`authorization.go` | 认证授权存储 |
-| `cluster_conf` | `cluster.go`、`sub_cluster.go`、`pool.go`、`api_key.go`、`model_price.go` | 集群与 API Key 存储 |
+| `api_key` | `api_key.go` | API-Key 存储 |
+| `cluster_conf` | `cluster.go`、`sub_cluster.go`、`pool.go`、`model_price.go` | 集群/子集群/实例池存储 |
+| `entity` | `entity.go`、`entity_type.go` | Entity / Entity-Type 存储 |
 | `route_conf` | `route_rule.go`、`domain.go` | 路由与域名存储 |
+| `route_rules` | `route_rules.go` | Global/Entity/API-Key 路由规则存储 |
 | `ai_route` | `ai_route.go` | AI 路由规则存储 |
 | `protocol` | `certificate.go` | 证书存储 |
-| `quota` | `entity.go`、`entity_type.go`、`quota_plan.go`、`quota_balance.go`、`rate_limit_policy.go`、`route_rules.go` | 配额与限流存储 |
+| `quota` | `quota_plan.go`、`quota_balance.go` | 配额计划/余额存储 |
+| `rate_limit_policy` | `rate_limit_policy.go` | 限流策略存储 |
 | `txn` | `txn.go` | 事务实现 |
 | `internal/dao` | `internal/curd.go`、`internal/builder.go`、各 `table_*.go` | 通用 DAO 与表映射 |
