+++ storage/rdb/api_key/api_key.go
@@ -12,7 +12,7 @@
 //See the License for the specific language governing permissions and
 //limitations under the License.
 
-package cluster_conf
+package api_key
 
 import (
 	"context"
