+++ storage/rdb/rate_limit_policy/rate_limit_policy.go
@@ -12,7 +12,7 @@
 //See the License for the specific language governing permissions and
 //limitations under the License.
 
-package quota
+package rate_limit_policy
 
 import (
 	"context"
