+++ storage/rdb/route_rules/route_rules.go
@@ -12,7 +12,7 @@
 //See the License for the specific language governing permissions and
 //limitations under the License.
 
-package quota
+package route_rules
 
 import (
 	"context"
